app/src/main/java/com/gemini/personalcapital/model/Article.java
package com.gemini.personalcapital.model;

import android.net.Uri;

public class Article {
    private String title;
    private String comments;
    private String pubDate;
    private String category;
    private String description;
    private String imageUrl;
    private String link;
    private Uri mCover;

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getComments() {
        return comments;
    }

    public void setComments(String comments) {
        this.comments = comments;
    }

    public String getPubDate() {
        return pubDate;
    }

    public void setPubDate(String pubDate) {
        this.pubDate = pubDate;
    }

    public String getCategory() {
        return category;
    }

    public void setCategory(String category) {
        this.category = category;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public String getImageUrl() {
        return imageUrl;
    }

    public void setImageUrl(String imageUrl) {
        this.imageUrl = imageUrl;
    }

    public Uri getmCover() {
        return mCover;
    }

    public void setmCover(Uri mCover) {
        this.mCover = mCover;
    }

    public String getLink() {
        return link;
    }

    public void setLink(String link) {
        this.link = link;
    }
}
